MOTECH-1901: Moved update of the UI changed fields
