+++ modules/tasks/tasks/src/main/java/org/motechproject/tasks/domain/Task.java
@@ -7,6 +7,7 @@ import org.motechproject.mds.annotations.Access;
 import org.motechproject.mds.annotations.Cascade;
 import org.motechproject.mds.annotations.Entity;
 import org.motechproject.mds.annotations.Field;
+import org.motechproject.mds.annotations.UIFilterable;
 import org.motechproject.mds.util.SecurityMode;
 import org.motechproject.tasks.constants.TasksRoles;
 import org.motechproject.tasks.json.TaskDeserializer;
@@ -52,6 +53,7 @@ public class Task {
     private TaskTriggerInformation trigger;
 
     @Field
+    @UIFilterable
     private boolean enabled;
 
     @Field
